Extract the `CumulativePageBucketReceiver` from `DistResultRXTask`.
This commits aims to decouple the `PageBucketReceiver` and the `DistResultRXTask`.
We will introduce buckets receivers that don't necessarily wait for all the
buckets of a page before processing the buckets and we want to be able to
plug these various strategies into the `DistResultRXTask`.
Also, this removes the `failure` method from `PageBucketReceiver` as `killed`
should always be used (there is no "resume" after a failure).
