+++ sql/src/main/java/io/crate/execution/engine/collect/collectors/RemoteCollector.java
@@ -29,7 +29,9 @@ import io.crate.data.RowConsumer;
 import io.crate.execution.dsl.phases.NodeOperation;
 import io.crate.execution.dsl.phases.RoutedCollectPhase;
 import io.crate.execution.engine.distribution.merge.PassThroughPagingIterator;
+import io.crate.execution.jobs.PageBucketReceiver;
 import io.crate.execution.jobs.DistResultRXTask;
+import io.crate.execution.jobs.CumulativePageBucketReceiver;
 import io.crate.execution.jobs.RootTask;
 import io.crate.execution.jobs.TasksService;
 import io.crate.execution.jobs.kill.KillJobsRequest;
@@ -168,15 +170,21 @@ public class RemoteCollector {
         } else {
             pagingIterator = PassThroughPagingIterator.oneShot();
         }
-        builder.addTask(new DistResultRXTask(
+        PageBucketReceiver pageBucketReceiver = new CumulativePageBucketReceiver(
             LOGGER,
             localNode,
             RECEIVER_PHASE_ID,
-            "RemoteCollectPhase",
             executor,
+            DataTypes.getStreamers(collectPhase.outputTypes()),
             consumer,
             pagingIterator,
-            DataTypes.getStreamers(collectPhase.outputTypes()),
+            1);
+
+        builder.addTask(new DistResultRXTask(
+            LOGGER,
+            RECEIVER_PHASE_ID,
+            "RemoteCollectPhase",
+            pageBucketReceiver,
             ramAccountingContext,
             1
         ));
